refactor: 구입 금액을 받는 메서드를 app에서 lottoMachine으로 이동
구입 금액과 관련된 로직은 이를 실질적으로 사용하는 lottoMachine에 더 적합하다 판단함
